finished

+++ libft/ft_striteri.c
@@ -6,7 +6,7 @@
 /*   By: yusuf <yusuf@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/13 13:15:15 by yciftci           #+#    #+#             */
-/*   Updated: 2024/03/07 19:26:54 by yusuf            ###   ########.fr       */
+/*   Updated: 2024/03/08 07:10:16 by yusuf            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	ft_striteri(char *s, void (*f)(unsigned int, char*))
 {
-	size_t	i;
+	int	i;
 
 	i = 0;
 	if (!s || !f)
